+++ R/summaries.R
@@ -17,12 +17,12 @@ Summary.units = function(..., na.rm = FALSE) {
       if (!ud.are.convertible(units(args[[i]]), u))
         stop(paste("argument", i, 
                    "has units that are not convertible to that of the first argument"))
-      args[[i]] = as.units(args[[i]], u) # convert to first unit
+      args[[i]] = as_units(args[[i]], u) # convert to first unit
     }
   }
   args = lapply(args, unclass)
-  # as.units(do.call(.Generic, args), u)
-  as.units(do.call(.Generic, c(args, na.rm = na.rm)), u)
+  # as_units(do.call(.Generic, args), u)
+  as_units(do.call(.Generic, c(args, na.rm = na.rm)), u)
 }
 
 #' @export